requirements.txt
fastapi
pydantic
uvicorn
sqlmodel
fastapi-mail
python-dotenv
email-validator